move fake server to integration tests
