+++ botstory/integrations/tests/__init__.py
@@ -1,0 +1,2 @@
+from .integration_test import *
+from .fake_server import *
