+++ botstory/integrations/tests/fake_server/__init__.py
@@ -1,0 +1,1 @@
+from .fake_fb import *
